storage: code movement in preparation of remap shard migration
