+++ src/storage-client/src/lib.rs
@@ -81,3 +81,4 @@ pub mod controller;
 pub mod sink;
 pub mod source;
 pub mod types;
+pub mod util;
