+++ src/storage/src/source/mod.rs
@@ -35,7 +35,6 @@ use crate::source::types::SourceMessageType;
 use crate::source::types::SourceReaderError;
 use crate::source::types::{NextMessage, SourceMessage, SourceReader};
 
-mod antichain;
 mod commit;
 mod delimited_value_reader;
 pub mod generator;
